server/requirements.txt
Django==3.2
djangorestframework
django-model-utils
django-import-export